move TICKET_TYPES to constants file
