+++ frontend/src/constants.js
@@ -1,3 +1,10 @@
+export const TICKET_TYPES = Object.freeze({
+    SF: 'SF',
+    FI: 'FI',
+    PPR: 'PPR',
+    UPR: 'UPR',
+})
+
 export const FACULTY_ADVISOR_EMAILS = ['drayside@uwaterloo.ca']
 export const TEAM_CAPTAIN_TITLES = ['Team Captain']
 export const DIRECTOR_TITLES = ['Admin', 'Director']
